requirements.txt
django
langchain
langchain-community
langchain-text-splitters
chromadb
pypdf
requests
python-dotenv
gunicorn
whitenoise
